Better startup procedure.
It is now possible to invoke the assemblage command line tool even if
there is no assemble.ml file in the cwd or access its manual even if
the parsing of assemble.ml fails or if the assemble.ml file doesn't
execute Assemblage.assemble.

We gathered all the command runtime environment handling in the As_env
module and added a -f option to specify the assemble.ml file to use.
As a result of this As_build_env.t becomes really only about the build
environment (it would before mix in things about the assemblage
startup procedure, -I, auto-load).

The command line help is adapted depending on whether we are using
assemblage the tool or directly executing an assemble.ml file
(e.g. the options about auto-load, -I etc. won't appear in the docs in
the latter case).